pyproject.toml
[tool.poetry]
name = "tweepy-lists"
version = "0.1.0"
description = ""
authors = ["Cody Bontecou <bontecouc@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.19.0"
tweepy = {git = "https://github.com/tweepy/tweepy.git"}
black = "^21.8b0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
